Abstracted bucket interactions to a separate class.
